microservices/requirements.txt
flask==2.3.3
flask-cors==5.0.0
pymongo[srv]==4.10.1
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.1.2 